+++ src/main/java/com/example/marimo_back/controller/ImageController.java
@@ -1,8 +1,7 @@
 package com.example.marimo_back.controller;
 
-import com.example.marimo_back.domain.ImageCollectionDto;
-import com.example.marimo_back.domain.ImageDto;
-import com.example.marimo_back.domain.Images;
+import com.example.marimo_back.Dto.ImageCollectionDto;
+import com.example.marimo_back.Dto.ImageDto;
 import com.example.marimo_back.service.ImageService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
